sensorTags subscription (only gen 1 atm)
API for easy connection and configuration (Period) to accelerometer and gyroscope. 
There is also a getter for the gattCharacteristiques if needed.

turnOnSensorTagServices just use the API to turn on with default parameter the two sensors, and subscribe to notification.

PS : make sure enable sensors at the right place in the code)